Add files via upload
Juzz Tell, an application developed using C++ and GTK for graphical user interface where you just tell the candidate requirements and it will tell you the perfect candidates for the job!It is an efficient interface that semantically matches and populates the best group of applicants from the numerous ones applied for the job by relating to the required fields and expectations of the organization. The different fields of basic qualifications and requirements and extra credits are given different weights and the applicants are ranked based on that thus filtering and populating the required and suitable group of applicants for the job.PDF and graph generation and multiple portal creation for the administrator, HR Manager and applicant are added features.